requirements.txt
annotated-types==0.7.0
anyio==4.3.0
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
Django==5.0.6
django-bootstrap5==24.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
openai==1.30.1
pillow==10.3.0
pydantic==2.7.1
pydantic_core==2.18.2
python-decouple==3.8
requests==2.32.1
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.0
tqdm==4.66.4
typing_extensions==4.11.0
urllib3==2.2.1
